Enabled i18nNext saveMissing with a script to catch the results
With `saveMissing` option enabled the i18nNext library sends POST
request to the mentioned endpoint with the missing locales.

We need this is because static parsers can not detect dynamic data we
are receiving from APIs.

During the development the server can be started using `node utility/missingLocaleServer.js`
The server will write missing locales to `missing_locales.json` file"

Then we need to replace public/locales/en.json file with
missing_locales/json.
The server is not directly editing public/locales/en.json file because
the create-react-app then keeps restarting the app in the development