Update App.scss
The position of the 'actions timeline' remains as it is even when you click on (or not, looking as if it is permanently set there '') ) language selection button. That portion of dates-timeline remains in its place whatsoever.
I first thought it is flex-box issue and was somewhat right. The thing is that is, it is a simple layout error due to wrong value assigned to position in 'header timeline'. The assigned value is set as 'absolute' but it is rather 'static'.
Assigning the value of position in header timeline as 'static' solves this issue.